test cpuflags
test cpuflags on x86_64 that the kernel is compatible with target cpu. thus if
you install rpm on 32bit booted kernel it could be still validated to be arch
compatible.

cpuflags check may be too aggressive, it could restrict installing stuff
to chroot if rpm implementation requires /proc to be mounted